Pretend we only want v2.1.0 of ui-plugin-find-eresource
In reality we need a newer version, but I need "release" v1.0.0 to be
able to build so I can publish it via the Jenkins pipeline, which I
need to do it's possible to install ui-plugin-eusage-reports from
folioci, whence it is ignored until the first release is published. We
live in stupid times.