Initial implementation of analyze-agreement button
This works, but is a bit ugly as we do not yet have information about
the present agreement passed in by ui-agreements. Consequently, we
extract the agreement ID from the URL (which works, but is rude), and
make do with not knowing its name (which makes the callouts less
informative).

Hopefully https://github.com/folio-org/ui-agreements/pull/828 will get
merged soon and we can do this properly.