Do not display old reportStatus when new one is a 404
This seems to be a bug in stripes-connect: once a resource has
successfully fetched data, a subsequent failed fetch (404 Not Found in
this case) results in the resource object's `records` element still
containing the records from the old, successful fetch.

Rather than leaping into stripes-connect and trying to fix this,
instead I am directly checking the `failed` element of the resource
object, and when it is set pretending that the record has not loaded,
thereby displaying "(undefined)" in the UI.

We may need to revisit this down the line if we need to distinguish
between "not yet known" because the resource hasn't loaded yet and
"definitely not set" because it's loaded but failed with a 404. But
since the former is a transitory state it doesn't seem worth worrying
about too much.