Use Stripes mocks from @folio/stripes-acq-components
This means I am able to dump the whole `test/jest/__mocks__`
directory, and use a well-maintained set of mocks instead of curating
my own.

The only downside is that for some reason the stripes-acq-components
mocks replace react-intl with a version that does not use translation
strings, so I have to unmock that in order to match against rendered
English text.