Rename report and picker labels
This involved adding a matrix of report-type x caption-type
translations to `en.json` and changing some existing translations.

Infuriatingly, the changed translations had to be copied across into
`en_US.json` otherwise the changes are obscured by the old versions in
the more specific locale file. Presumably someone will have to do the
same for the other English locales. This is silly: it would be better
if the specific locales included _only_ the keys for which they
provide a different value from that in `en.json`. /shrug/

There's a lot of xCaption and yCaption parameter passing into the
components for the individual graphs, and through their corresponding
loaders. There's probably a "better" way, but it would likely be
harder to read, so swings and roundabouts.

Fixes UIPER-76.