Matching summary uses facets for matched/unmatched/etc. counts
We now fetch with limit=0 to avoid fetching any actual records because
all we want for the summary is the counts in the facets.

As a result, we now display correct counts rather than counts trucated
to at most 1000 records. But the match editor does not work _at all_
since it is not fetching its own data.

Processing the facets into the form we want to use involves a new
utility function, so I wrote tests for it.