Implement switching between tabs in the match editor
This works by generating a `query` parameter in the stripes-connect
resource's `params` function. (At present this does not also take into
account a user-entered query, but that's not a problem as full
searching, sorting and filtering lies outside the scope of UIPER-16.)

**NOTE.** Although this results in sending what I think are the
correct queries, and this works for the Loaded and Matched tab,
presently the Unmatched and Ignored tabs cannot be displayed due to
what I think is a bug in the back-end: see MODEUR-106.