Add "disableAutosize" prop to <MatchEditor>
By exhaustive and painful experiment, I have determined that when
rendered by RTL, the actual content of a `<MultiColumnList>` is not
included in the render if the `autosize` attribute is true.

So in order to make this test work, I had to add a "disableAutosize"
attribute which the test-script can pass in. In other words, I had to
add a complication to my code that is not needed at all when it's
actually running.

_sigh_