Include <FormattedMessage> italic-callbacks in coverage
It turns out that they _can_ be forced to run, but it takes some work
in the testing script.

What's returned from `<FormattedMessage>` is not a formatted message
-- oh, no! -- but a "message object", which needs to be explicitly
resolved to an actual message by `ReactDOMServer.renderToString`. Only
then are formatting callbacks like `i: x => <i>{x}</i>` executed.

This is painful to do. But the upside is that it makes it possible for
to assert on the results of translations/substitutions. So we do
actually get more meaningful tests out of it.